% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_p.R, R/as_d.R, R/as_q.R, R/as_r.R,
%   R/utils-as.R
\name{as_p}
\alias{as_p}
\alias{as_p.default}
\alias{as_p.pdqr}
\alias{as_d}
\alias{as_d.default}
\alias{as_d.pdqr}
\alias{as_q}
\alias{as_q.default}
\alias{as_q.pdqr}
\alias{as_r}
\alias{as_r.default}
\alias{as_r.pdqr}
\alias{as-pdqr}
\title{Convert to pdqr-function}
\usage{
as_p(f, ...)

\method{as_p}{default}(f, support = NULL, ..., n_grid = 10001)

\method{as_p}{pdqr}(f, ...)

as_d(f, ...)

\method{as_d}{default}(f, support = NULL, ..., n_grid = 10001)

\method{as_d}{pdqr}(f, ...)

as_q(f, ...)

\method{as_q}{default}(f, support = NULL, ..., n_grid = 10001)

\method{as_q}{pdqr}(f, ...)

as_r(f, ...)

\method{as_r}{default}(f, support = NULL, ..., n_grid = 10001,
  n_sample = 10000, args_new = list())

\method{as_r}{pdqr}(f, ...)
}
\arguments{
\item{f}{Appropriate function to be converted (see Details).}

\item{...}{Extra arguments for \code{f}.}

\item{support}{Support of output. If \code{NULL} or any value is \code{NA}, it is
detected using specific algorithms (see Details).}

\item{n_grid}{Number of grid points at which \code{f} will be evaluated (see
Details).}

\item{n_sample}{Number of points to sample from \code{f}.}

\item{args_new}{List of extra arguments for \code{\link[=new_d]{new_d()}} to control \code{\link[=density]{density()}}.}
}
\value{
A pdqr-function of corresponding class.
}
\description{
Convert some function to be a proper pdqr-function of type "infin".
}
\examples{
# Convert existing "proper" pdqr-function
set.seed(101)
x <- rnorm(10)
my_d <- new_d(x, "infin")

my_p <- as_p(my_d)

# Convert some other function to be a "proper" pdqr-function
p_unif <- as_p(punif)
r_beta <- as_r(rbeta, shape1 = 2, shape2 = 2)

}
