% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form-other.R
\name{form_mix}
\alias{form_mix}
\title{Construct mixture of distributions}
\usage{
form_mix(f_list, weights = NULL)
}
\arguments{
\item{f_list}{List of pdqr-functions.}

\item{weights}{Numeric vector of weights.}
}
\value{
A pdqr-function for mixture distribution.
}
\description{
Based on a list of pdqr-functions and vector of weights construct a
pdqr-function of corresponding mixture distribution.
}
\examples{
d_unif <- as_d(dunif)
p_norm <- as_p(pnorm)

my_mix <- form_mix(list(d_unif, p_norm), weights = c(0.7, 0.3))
plot(my_mix)

}
