% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form-other.R
\name{form_smooth}
\alias{form_smooth}
\title{Smooth pdqr-function}
\usage{
form_smooth(f, n_sample = 10000, args_new = list())
}
\arguments{
\item{f}{Pdqr-function.}

\item{n_sample}{Number of elements to sample.}

\item{args_new}{List of extra arguments for \link[=new_d]{new_*()} to control
\code{\link[=density]{density()}}.}
}
\value{
A smoothed version of \code{f}.
}
\description{
Smooth pdqr-function using random sampling and corresponding
\link[=new_d]{new_*()} function.
}
\examples{
set.seed(101)

# Pdqr-functions of type "infin"
bad_infin <- new_d(data.frame(x = sort(runif(100)), y = runif(100)), "infin")
smoothed_infin <- form_smooth(bad_infin)
plot(bad_infin)
lines(smoothed_infin, col = "blue")

# Pdqr-functions of type "fin"
bad_fin <- new_d(data.frame(x = sort(runif(100)), prob = runif(100)), "fin")
smoothed_fin <- form_smooth(bad_fin)
plot(bad_fin)
lines(smoothed_fin, col = "blue")

}
