% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_tails.R
\name{form_tails}
\alias{form_tails}
\title{Transform tails of distribution}
\usage{
form_tails(f, level, method = "trim", direction = "both")
}
\arguments{
\item{f}{A pdqr-function.}

\item{level}{Cutoff level.}

\item{method}{Transformation method. One of "trim" or "winsor".}

\item{direction}{Information about which tail(s) to transform. One of "both",
"left", or "right".}
}
\value{
A pdqr-function with transformed tail(s).
}
\description{
Modify tail(s) of distribution defined by certain cutoff level using method
of choice.
}
\examples{
d_norm <- as_d(dnorm)
plot(d_norm)
lines(form_tails(d_norm, level = 0.1), col = "blue")
lines(
  form_tails(d_norm, level = 0.1, method = "winsor", direction = "right"),
  col = "green"
)

# Also works with pdqr-functions of type "fin"
my_fin <- new_d(data.frame(x = 1:4, prob = (1:4)/10), type = "fin")
meta_x_tbl(form_tails(my_fin, level = 0.1))
meta_x_tbl(
  form_tails(my_fin, level = 0.35, method = "winsor", direction = "left")
)

}
