% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta}
\alias{meta}
\alias{meta_all}
\alias{meta_type}
\alias{meta_support}
\alias{meta_x_tbl}
\title{Get metadata of pdqr-functions}
\usage{
meta_all(f)

meta_type(f)

meta_support(f)

meta_x_tbl(f)
}
\arguments{
\item{f}{A "proper" pdqr-function.}
}
\value{
\code{meta_all()} returns a list of all metadata. \code{meta_type()},
\code{meta_support}, and \code{meta_x_tbl()} return corresponding metadata.
}
\description{
Get metadata of pdqr-functions
}
\examples{
d_unif <- as_d(dunif)

meta_all(d_unif)
meta_type(d_unif)
meta_support(d_unif)
head(meta_x_tbl(d_unif))

}
