% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_p.R, R/new_d.R, R/new_q.R, R/new_r.R,
%   R/utils-new.R
\name{new_p}
\alias{new_p}
\alias{new_d}
\alias{new_q}
\alias{new_r}
\alias{new-pdqr}
\title{Create new pdqr-function}
\usage{
new_p(x, type = "infin", ...)

new_d(x, type = "infin", ...)

new_q(x, type = "infin", ...)

new_r(x, type = "infin", ...)
}
\arguments{
\item{x}{Numeric vector or data frame with appropriate columns (see Details).}

\item{type}{Type of pdqr-function. One of "fin" or "infin" (see Details).}

\item{...}{Extra arguments for \link[stats:density]{density()}.}
}
\value{
A pdqr-function of corresponding class.
}
\description{
Main tools to create new distribution pdqr-functions based on numeric sample
or data frame describing distribution.
}
\examples{
set.seed(101)
x <- rnorm(10)
my_d_infin <- new_d(x, "infin")
my_d_fin <- new_d(x, "fin")

my_p_fin <- new_p(data.frame(x = 1:10, prob = rep(1, 10)/10), "fin")

# Using bigger bandwidth in `density()`
my_d_infin_2 <- new_d(x, "infin", adjust = 2)

}
