% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ-other.R
\name{summ-prob}
\alias{summ-prob}
\alias{summ_prob_false}
\alias{summ_prob_true}
\title{Summarize boolean distribution with probability}
\usage{
summ_prob_false(f)

summ_prob_true(f)
}
\arguments{
\item{f}{A pdqr-function representing distribution.}
}
\value{
A single numeric value representing corresponding probability.
}
\description{
Here \code{summ_prob_false()} returns a probability of 0 and \code{summ_prob_true()} -
complementary probability (one minus \code{summ_prob_false()} output). Both of
them check if their input is a \strong{boolean} pdqr-function: type "fin" with \code{x}
in \code{x_tbl} identical to \code{c(0, 1)}. If it is not, warning is thrown.
}
\examples{
d_unif <- as_d(dunif)
d_norm <- as_d(dnorm)
summ_prob_true(d_unif > d_norm)
summ_prob_false(2*d_norm > d_unif)

# When input is "infin" function or doesn't have 0 as distribution element,
# probability of being false is returned as 0.
summ_prob_false(d_unif)
summ_prob_true(new_d(2, "fin"))

}
