% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_dispersion.R
\name{summ_dispersion}
\alias{summ_dispersion}
\alias{summ_sd}
\alias{summ_var}
\alias{summ_iqr}
\alias{summ_mad}
\title{Summarize distribution with dispersion}
\usage{
summ_dispersion(f, method = "sd")

summ_sd(f)

summ_var(f)

summ_iqr(f)

summ_mad(f)
}
\arguments{
\item{f}{A pdqr-function representing distribution.}

\item{method}{Method of center computation. Should be one of "sd", "var",
"iqr", "mad".}
}
\value{
All functions return a single number representing a dispersion of
distribution.
}
\description{
Functions to compute dispersion of distribution (i.e. "how wide it is
spread"). \code{summ_dispersion()} is a wrapper for respective \code{summ_*()}
functions (from this page) with default arguments.
}
\details{
\code{summ_sd()} computes distribution's standard deviation.

\code{summ_var()} computes distribution's variance.

\code{summ_iqr()} computes distribution's interquartile range. Essentially, it is
a \code{as_q(f)(0.75) - as_q(f)(0.25)}.

\code{summ_mad()} computes distribution's \emph{median} absolute deviation around the
distribution's \emph{median}.
}
\examples{
# Type "infin"
d_norm <- as_d(dnorm)
  # The same as `summ_dispersion(d_norm, method = "sd")`
summ_sd(d_norm)
summ_var(d_norm)
summ_iqr(d_norm)
summ_mad(d_norm)

# Type "fin"
d_pois <- as_d(dpois, lambda = 10)
summ_sd(d_pois)
summ_var(d_pois)
summ_iqr(d_pois)
summ_mad(d_pois)

}
\seealso{
\code{\link[=summ_center]{summ_center()}} for computing distribution's center, \code{\link[=summ_moment]{summ_moment()}}
for general moments.

Other summary functions: \code{\link{summ_center}},
  \code{\link{summ_distance}}, \code{\link{summ_entropy}},
  \code{\link{summ_hdr}}, \code{\link{summ_interval}},
  \code{\link{summ_moment}}, \code{\link{summ_prob_true}},
  \code{\link{summ_pval}}, \code{\link{summ_quantile}},
  \code{\link{summ_separation}}
}
\concept{summary functions}
