% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ-other.R
\name{summ_quantile}
\alias{summ_quantile}
\title{Compute quantiles of distribution}
\usage{
summ_quantile(f, probs)
}
\arguments{
\item{f}{A pdqr-function representing distribution.}

\item{probs}{Vector of probabilities for which quantiles should be returned.}
}
\value{
A numeric vector of the same length as \code{probs} representing
corresponding quantiles.
}
\description{
Essentially, this is a more strict wrapper of \code{as_q(f)(probs)}. If any value
in \code{probs} is outside of segment \[0; 1\], an error is thrown.
}
\examples{
d_norm <- as_d(dnorm)
probs <- c(0.25, 0.5, 0.75)
all.equal(summ_quantile(d_norm, probs), as_q(d_norm)(probs))

}
