% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_retype.R
\name{form_retype}
\alias{form_retype}
\title{Change type of pdqr-function}
\usage{
form_retype(f, type, method = "piecelin")
}
\arguments{
\item{f}{A pdqr-function.}

\item{type}{A desired type of output. Should be one of "fin" or "infin".}

\item{method}{Retyping method. Should be one of "piecelin" or "dirac".}
}
\value{
A pdqr-function with type equal to input \code{type}.
}
\description{
Modify \link[=meta_type]{type} of pdqr-function using method of choice.
}
\details{
If type of \code{f} is equal to input \code{type} then \code{f} is returned.

Method "piecelin" (default) should be used mostly for converting from "infin"
to "fin" type. It uses the fact that 'pdqr' densities are piecewise-linear
(linear in intervals between values of "x" column of \link[=meta_x_tbl]{"x_tbl" metadata}) on their \link[=meta_support]{support}:
\itemize{
\item Retyping from "infin" to \code{type} "fin" is done by computing "x" values as
centers of interval masses with probabilities equal to interval total
probabilities.
\item Retyping from "fin" to \code{type} "infin" is made approximately by trying to
compute "x" grid, for which "x" values of input distribution are going to be
centers of mass. Algorithm is approximate and might result into a big errors
in case of small number of "x" values or if they are not "suitable" for this
kind of transformation.
}

Method "dirac" is used mostly for converting from "fin" to "infin" type (for
example, in \code{form_mix()} in case different types of input pdqr-functions). It
works in the following way:
\itemize{
\item Retyping from "infin" to \code{type} "fin" works only if "x_tbl" metadata
represents a mixture of dirac-like distributions. In that case it is
transformed to have "x" values from centers of those dirac-like distributions
with corresponding probabilities.
\item Retyping from "fin" to \code{type} "infin" works by transforming each "x" value
from "x_tbl" metadata into dirac-like distribution with total probability
taken from corresponding value of "prob" column. Output essentially
represents a mixture of dirac-like distributions.
}
}
\examples{
my_infin <- new_d(data.frame(x = 1:5, y = c(1, 2, 3, 2, 1)/9), "infin")
my_fin <- form_retype(my_infin, "fin")
meta_x_tbl(my_fin)

my_dirac <- form_retype(my_fin, "infin", method = "dirac")
meta_x_tbl(my_dirac)

# Default retyping from "fin" to "infin" isn't very accurate for small number
# of points/intervals
my_infin_2 <- form_retype(my_fin, "infin")
meta_x_tbl(my_infin_2)

plot(my_infin)
lines(my_infin_2, col = "blue")

}
\seealso{
\code{\link[=form_regrid]{form_regrid()}} for changing grid (rows of "x_tbl" metadata) of
pdqr-function.

\code{\link[=form_resupport]{form_resupport()}} for changing support of pdqr-function.

Other form functions: \code{\link{form_estimate}},
  \code{\link{form_mix}}, \code{\link{form_regrid}},
  \code{\link{form_resupport}}, \code{\link{form_smooth}},
  \code{\link{form_tails}}, \code{\link{form_trans}}
}
\concept{form functions}
