% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_retype.R
\name{form_retype}
\alias{form_retype}
\title{Change type of pdqr-function}
\usage{
form_retype(f, type, method = "piecelin")
}
\arguments{
\item{f}{A pdqr-function.}

\item{type}{A desired type of output.}

\item{method}{Retyping method.}
}
\value{
A pdqr-function with type \code{type}.
}
\description{
Modify type of pdqr-function ("infin" or "fin") using method of choice.
}
\examples{
my_infin <- new_d(data.frame(x = 1:5, y = c(1, 2, 3, 2, 1)/9), "infin")
my_fin <- form_retype(my_infin, "fin")
meta_x_tbl(my_fin)

my_dirac <- form_retype(my_fin, "infin", method = "dirac")
meta_x_tbl(my_dirac)

# Default retyping from "fin" to "infin" isn't very accurate for small number
# of points/intervals
my_infin_2 <- form_retype(my_fin, "infin")
meta_x_tbl(my_infin_2)

plot(my_infin)
lines(my_infin_2, col = "blue")

}
